Created models from existing types.ts; updated refs
