+++ src/components/OptionDetailsView.tsx
@@ -6,7 +6,7 @@ import { Button, IconButton } from "@mui/material";
 import { ethers } from "ethers";
 import { account, signer } from "../utils/metamask";
 import { getAccountDisplayValue, dismissLastToast, showToast, flavorLabels } from "../utils/frontend";
-import {  OptionState, OptionWithAsset } from "../utils/types";
+import {  OptionState, OptionWithAsset } from "../models/option";
 import { ADDRESS0 } from "../utils/constants";
 import { useEffect, useState } from "react";
 import toast from "react-hot-toast";
