+++ src/components/OptionRequestForm.tsx
@@ -3,9 +3,9 @@ import classes from "../styles/components/OptionRequestForm.module.scss";
 
 import { Button, FormControl, RadioGroup } from "@mui/material";
 import clsx from "clsx";
-import { useRef, useState } from "react";
+import { useState } from "react";
 import { BIGNUMBER0, SECONDS_IN_A_DAY } from "../utils/constants";
-import { NFTAsset, OptionFlavor, OptionRequest } from "../utils/types";
+import { OptionFlavor } from "../models/option";
 import { ethers } from "ethers";
 import { loadNFTImage, imageOf, keyOf } from "../datasources/NFT/localhost";
 import { getFloatString, getIntervalString, showToast } from "../utils/frontend";
@@ -13,6 +13,8 @@ import TextBox_OptionRequestForm from "../fragments/TextBox.OptionRequest";
 import Radio_OptionRequestForm from "../fragments/Radio.OptionRequest";
 import DropDown_OptionRequestForm from "../fragments/DropDown.OptionRequest";
 import { useContracts } from "../pages/_app";
+import { OptionRequest } from "../models/optionRequest";
+import { NFTAsset } from "../models/nftAsset";
 
 const createRequest = (obj : OptionRequest) =>
 {
