+++ src/models/nftAsset.ts
@@ -1,0 +1,9 @@
+import { BigNumber } from "ethers";
+
+export type NFTAsset =
+{
+    nftId       : BigNumber
+,   nftContract : string
+,   name        : string
+,   image       : string
+};
