+++ test/withdrawOptionRequest.ts
@@ -1,5 +1,5 @@
 import { expect } from "chai";
-import { OptionState } from "../src/utils/types";
+import { OptionState } from "../src/models/option";
 
 import { buyer, seller, initializer, dummyOptionRequest, publishDummyOptionRequest } from "./utils/helpers";
 import { NFTOptContract, deployMainContract } from "../src/utils/deployment";
